refactor: separate out prose editor
